Conflicts:
	src/main/NGECore.java
	src/services/DevService.java
	src/services/command/CommandService.java

+++ src/services/PlayerService.java
@@ -64,6 +64,7 @@ import resources.common.RGB;
 import resources.common.SpawnPoint;
 import resources.common.StringUtilities;
 import resources.datatables.PlayerFlags;
+import resources.datatables.Professions;
 import resources.guild.Guild;
 import resources.objects.Buff;
 import resources.objects.building.BuildingObject;
@@ -123,21 +124,17 @@ public class PlayerService implements INetworkDispatch {
 		}, 30, 30, TimeUnit.SECONDS);
 		
 		scheduler.scheduleAtFixedRate(() -> {
-			synchronized(creature.getMutex()) {
-				if(creature.getAction() < creature.getMaxAction() && creature.getPosture() != 14) {
-					if(creature.getCombatFlag() == 0)
-						creature.setAction(creature.getAction() + (15 + creature.getLevel() * 5));
-					else
-						creature.setAction(creature.getAction() + ((15 + creature.getLevel() * 5) / 2));
-				}
+			if(creature.getAction() < creature.getMaxAction() && creature.getPosture() != 14) {
+				if(creature.getCombatFlag() == 0)
+					creature.setAction(creature.getAction() + (15 + creature.getLevel() * 5));
+				else
+					creature.setAction(creature.getAction() + ((15 + creature.getLevel() * 5) / 2));
 			}
 		}, 0, 1000, TimeUnit.MILLISECONDS);
 
 		scheduler.scheduleAtFixedRate(() -> {
-			synchronized(creature.getMutex()) {
-				if(creature.getHealth() < creature.getMaxHealth() && creature.getCombatFlag() == 0 && creature.getPosture() != 13 && creature.getPosture() != 14)
-					creature.setHealth(creature.getHealth() + (36 + creature.getLevel() * 4));
-			}
+			if(creature.getHealth() < creature.getMaxHealth() && creature.getCombatFlag() == 0 && creature.getPosture() != 13 && creature.getPosture() != 14)
+				creature.setHealth(creature.getHealth() + (36 + creature.getLevel() * 4));
 		}, 0, 1000, TimeUnit.MILLISECONDS);
 		
 		/*final PlayerObject ghost = (PlayerObject) creature.getSlottedObject("ghost");
@@ -503,12 +500,13 @@ public class PlayerService implements INetworkDispatch {
 		if(cell == null)
 			return;
 		
+		creature.setPosture((byte) 0);
+		
 		core.simulationService.transferToPlanet(creature, cloner.getPlanet(), spawnPoint.getPosition(), spawnPoint.getOrientation(), cell);
 		
 		creature.setHealth(creature.getMaxHealth());
 		creature.setAction(creature.getMaxAction());
 		
-		creature.setPosture((byte) 0);
 		creature.setSpeedMultiplierBase(1);
 		creature.setTurnRadius(1);
 		
@@ -536,8 +534,8 @@ public class PlayerService implements INetworkDispatch {
 		}
 		
 		player.setProfession(profession);
-		
-		String xpType = ((player.getProfession().contains("entertainer")) ? "entertainer" : ((player.getProfession().contains("trader")) ? "crafting" : "combat_general"));
+
+		String xpType = ((profession.contains("entertainer")) ? "entertainer" : ((profession.contains("trader")) ? "crafting" : "combat_general"));
 			
 		int experience = player.getXp(xpType);
 		
@@ -547,7 +545,7 @@ public class PlayerService implements INetworkDispatch {
 			for (int i = 0; i < experienceTable.getRowCount(); i++) {
 				if (experienceTable.getObject(i, 0) != null) {
 					if (experience >= ((Integer) experienceTable.getObject(i, 1))) {
-						level = (Integer) experienceTable.getObject(i, 1);
+						level = (Integer) experienceTable.getObject(i, 0);
 					}
 				}
 			}
@@ -556,10 +554,11 @@ public class PlayerService implements INetworkDispatch {
 		}
 		
 		grantLevel(creature, level);
+		player.setProfessionIcon(Professions.get(profession));
 	}
 	
 	/*
-	 * Resets to level 0
+	 * Resets to level 1
 	 */
 	public void resetLevel(CreatureObject creature) {
 		PlayerObject player = (PlayerObject) creature.getSlottedObject("ghost");
@@ -577,18 +576,13 @@ public class PlayerService implements INetworkDispatch {
         				case "object/tangible/datapad/shared_character_datapad.iff":
         				case "object/tangible/bank/shared_character_bank.iff":
         				case "object/tangible/mission_bag/shared_mission_bag.iff":
-        				case "object/weapon/creature/shared_creature_default_weapon.iff": {
+        				case "object/weapon/creature/shared_creature_default_weapon.iff":
         					continue;
-        				}
-        				default: {
-        					//
-        				}
+        				default:
+        					core.equipmentService.unequip(creature, equipment);
         			}
-        			
-        			core.equipmentService.unequip(creature, equipment);
         		}
-		}
-		catch (Exception e) {
+		} catch (Exception e) {
 			e.printStackTrace();
 		}
 		
@@ -598,9 +592,31 @@ public class PlayerService implements INetworkDispatch {
 		
 		core.buffService.clearBuffs(creature);
 		
-		for (String skill : creature.getSkills()) {
-			core.skillService.removeSkill(creature, skill);
+		try {
+			String[] skills;
+			
+			DatatableVisitor skillTemplate = ClientFileManager.loadFile("datatables/skill_template/skill_template.iff", DatatableVisitor.class);
+			
+			for (int s = 0; s < skillTemplate.getRowCount(); s++) {
+				if (skillTemplate.getObject(s, 0) != null) {
+					if (((String) skillTemplate.getObject(s, 0)).equals(player.getProfession())) {
+						skills = ((String) skillTemplate.getObject(s, 4)).split(",");
+						
+						for (String skill : skills) {
+							core.skillService.removeSkill(creature, skill);
+						}
+						
+						core.skillService.addSkill(creature, skills[0]);
+						
+						break;
+					}
+				}
+			}
+		}  catch (InstantiationException | IllegalAccessException e) {
+			e.printStackTrace();
 		}
+		
+		core.skillService.resetExpertise(creature);
 		
 		String xpType = ((player.getProfession().contains("entertainer")) ? "entertainer" : ((player.getProfession().contains("trader")) ? "crafting" : "combat_general"));
 			
@@ -620,7 +636,7 @@ public class PlayerService implements INetworkDispatch {
 		creature.setAction(300);
 		creature.setGrantedHealth(0);
 		
-		creature.setLevel((short) 0);
+		creature.setLevel((short) 1);
 	}
 	
 	/*
@@ -638,11 +654,9 @@ public class PlayerService implements INetworkDispatch {
 			return;
 		}
 		
-		resetLevel(creature);
+		if(level == 0) return;
 		
-		if (level == 0) {
-			return;
-		}
+		resetLevel(creature);
 		
 		try {
 			experienceTable = ClientFileManager.loadFile("datatables/player/player_level.iff", DatatableVisitor.class);
@@ -803,6 +817,11 @@ public class PlayerService implements INetworkDispatch {
 		DatatableVisitor experienceTable;
 		PlayerObject player = (PlayerObject) creature.getSlottedObject("ghost");
 		experience *= xpMultiplier;
+		
+		if (creature.getLevel() >= 90) {
+			return;
+		}
+		
 		//synchronized(objectMutex) {
 			try {
 				experienceTable = ClientFileManager.loadFile("datatables/player/player_level.iff", DatatableVisitor.class);
@@ -815,7 +834,7 @@ public class PlayerService implements INetworkDispatch {
 				experience += ((experience * experienceBonus) / 100);
 				
 				// 1. Add the experience.
-				if (experience > 0) {
+				if (experience > 0 && !creature.isStationary()) {
 					creature.showFlyText("base_player", "prose_flytext_xp", "", experience, (float) 2.5, new RGB(180, 60, 240), 1);
 				}
 				
@@ -1207,6 +1226,25 @@ public class PlayerService implements INetworkDispatch {
 			}
 		});
 		core.suiService.openSUIWindow(ringWindow);
+	}
+	
+	public String getFormalProfessionName(String template) {
+		String formalName = "";
+
+		switch (template) {
+		case "force_sensitive_1a":	formalName = "Jedi"; break;
+		case "bounty_hunter_1a":	formalName = "Bounty Hunter"; break;
+		case "officer_1a":			formalName = "Officer"; break;
+		case "smuggler_1a":			formalName = "Smuggler"; break;
+		case "entertainer_1a":		formalName = "Entertainer"; break;
+		case "spy_1a":				formalName = "Spy"; break;
+		case "medic_1a":			formalName = "Medic"; break;
+		case "commando_1a":			formalName = "Commando"; break;
+		
+		default:					formalName = "Trader"; break;	// Ziggy: Trader profession names are a bit irregular, so this is used.
+
+		}
+		return formalName;
 	}
 	
 	@Override
